pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["nicolevito <coemakeup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
factory-boy = "^3.3.3"
pytest = "^8.3.5"
django = "4.2"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
